pc/at refactored computestats for issue timeline events.
