+++ app/javascript/graphql/IssueTimelineItems.js
@@ -39,6 +39,33 @@ class IssueTimelineItems extends GithubGraphqlQuery {
 
         `); /* EndGraphQL */
     }
+    static computeStats (data, databaseId_to_team){
+        let statistics = {};
+        let errors = {};
+
+        try {
+            let issues = data.data.repository.issues;
+            let issueNodes = issues.nodes;
+            let timelineItemsTotalCountVector =
+               issueNodes.map( (n) => n.timelineItems.totalCount);
+            let sum = (a,b)=>a+b;
+            let timelineItemsCount = 
+                timelineItemsTotalCountVector.reduce(sum, 0)
+            statistics["totalIssues"] = issues.totalCount;
+            statistics["timelineItemsCount"] = timelineItemsCount;
+        } catch(e) { 
+             errors = {
+                 name : e.name,
+                 message: e.message
+             };
+         }
+
+        return {
+            statistics: statistics,
+            errors: errors
+        };
+        return {}
+    }
 }
 
 export default IssueTimelineItems;
